Add inheritance for different build managers
